switch to push op, fix sugg. sorting, reroute to auth page, fix like pushing bug, comments look better, mobile reactive fix

+++ app/b/[board_name]/page.tsx
@@ -11,6 +11,7 @@ import { useUser } from "@clerk/nextjs"
 import { toast } from "react-toastify"
 import "react-toastify/dist/ReactToastify.css"
 import Image from "next/image"
+import { SignInToastMessage } from "@/components/board/SignInToastMessage"
 
 export default function BoardInfo({ params }: { params: { board_name: string } }) {
   const [error, setError] = useState<string | null>(null)
@@ -28,7 +29,7 @@ export default function BoardInfo({ params }: { params: { board_name: string } }
   useEffect(() => {
     if (board) {
       // set tab title
-      document.title = board.name
+      document.title = `Feedback | ${board.name}`
       // set favicon
       const favicon = (document.querySelector("link[rel='icon']") as HTMLLinkElement) || document.createElement("link")
       favicon.type = "image/x-icon"
@@ -82,7 +83,6 @@ export default function BoardInfo({ params }: { params: { board_name: string } }
       userLikes = board.suggestions
         .filter((suggestion: Suggestion) => suggestion.votes.some((vote: Vote) => vote.author === user.id))
         .map((suggestion: Suggestion) => suggestion.id)
-      console.log(userLikes)
       if (existingUser.id) {
         // if they already have an entry in local storage
         const updatedLikes = [...existingUser.likedSuggestions, ...userLikes]
@@ -99,16 +99,18 @@ export default function BoardInfo({ params }: { params: { board_name: string } }
         }
         localStorage.setItem("user", JSON.stringify(newUser))
       }
-      // } else if (!isSignedIn && existingUser.id) {
-      //   // not logged in but have visited before
-      //   return
     } else if (!isSignedIn && isLoaded && board) {
-      // not logged in and first visit OR logged in user that logged out
-      const newUser: LocalStorageUser = {
-        id: uuidv4(),
-        likedSuggestions: [],
+      if (existingUser.id) {
+        // not logged in and not first visit or logged in user that logged out
+        return // no action needs to be taken
+      } else {
+        // not logged in and first visit
+        const newUser: LocalStorageUser = {
+          id: uuidv4(),
+          likedSuggestions: [],
+        }
+        localStorage.setItem("user", JSON.stringify(newUser))
       }
-      localStorage.setItem("user", JSON.stringify(newUser))
     }
   }
 
@@ -121,16 +123,22 @@ export default function BoardInfo({ params }: { params: { board_name: string } }
   }
 
   const handleNewSuggestionSubmission = async () => {
-    if (!suggestionTitle || !suggestionDescription) return
+    const trimmedTitle = suggestionTitle.trim()
+    const trimmedDescription = suggestionDescription.trim()
 
-    if (!isSignedIn) return // TODO: Add sign in modal
+    if (!trimmedTitle || !trimmedDescription) return
 
-    if (suggestionTitle.length > 250) {
+    if (!isSignedIn) {
+      SignInToastMessage()
+      return
+    }
+
+    if (trimmedTitle.length > 250) {
       toast.error("Suggestion title must be less than 250 characters")
       return
     }
 
-    if (suggestionDescription.length > 500) {
+    if (trimmedDescription.length > 500) {
       toast.error("Suggestion description must be less than 500 characters")
       return
     }
@@ -143,12 +151,12 @@ export default function BoardInfo({ params }: { params: { board_name: string } }
         headers: {
           "Content-Type": "application/json",
         },
-        body: JSON.stringify({ title: suggestionTitle, description: suggestionDescription, board: board }),
+        body: JSON.stringify({ title: trimmedTitle, description: trimmedDescription, board: board }),
       })
 
       if (!response.ok) {
         const errorData = await response.json()
-        throw new Error(errorData.message || "Failed to add comment")
+        throw new Error(errorData.message || "Failed to post feedback")
       } else {
         const data = await response.json()
         setBoard((prevBoard) => {
@@ -185,7 +193,7 @@ export default function BoardInfo({ params }: { params: { board_name: string } }
 
       if (!response.ok) {
         const errorData = await response.json()
-        throw new Error(errorData.message || "Failed to add comment")
+        throw new Error(errorData.message || "Failed to load more suggestions")
       } else {
         const data = await response.json()
         setPage(page + 1)
@@ -210,72 +218,70 @@ export default function BoardInfo({ params }: { params: { board_name: string } }
 
   return (
     <main
-      className="flex flex-col items-center min-h-screen w-full"
+      className="flex md:flex-row flex-col max-w-7xl md:p-4 p-0 mx-auto min-h-screen w-full"
       style={{ backgroundColor: board?.primaryColor || "#fff", color: board?.textColor || "#000" }}
     >
-      <div className="w-full max-w-7xl mx-auto p-4 flex">
-        <div className="w-1/3 p-4">
-          <div>
-            {board?.logo ? (
-              <Image src={board.logo} alt={`${board.name} logo`} width={175} height={175} className="mb-4" />
-            ) : (
-              <p className="text-xl font-bold mb-4 w-full break-words">{board?.name}</p>
-            )}
-          </div>
-          <div className="p-6 rounded-lg" style={{ backgroundColor: board?.secondaryColor || "#f9fafb" }}>
-            <h1 className="text-lg font-semibold mb-6">Add feedback</h1>
-            <div className="mb-4">
-              <label className="block text-xs font-bold mb-1" htmlFor="suggestionTitle">
-                Post Title
-              </label>
-              <input
-                id="suggestionTitle"
-                type="text"
-                placeholder="Add magic link authentication"
-                className="w-full p-2 rounded-lg focus:outline-none"
-                value={suggestionTitle}
-                onChange={(e) => setsuggestionTitle(e.target.value)}
-                style={{ backgroundColor: lighterSecondaryColor }}
-              />
-            </div>
-            <div className="mb-4">
-              <label className="block text-xs font-bold mb-1" htmlFor="suggestionDescription">
-                Post Text
-              </label>
-              <textarea
-                id="suggestionDescription"
-                placeholder="I hate having to keep track of all these passwords..."
-                className="w-full p-2 rounded-lg focus:outline-none"
-                value={suggestionDescription}
-                onChange={(e) => setsuggestionDescription(e.target.value)}
-                style={{ backgroundColor: lighterSecondaryColor, height: "100px" }}
-              />
-            </div>
-            <button
-              onClick={handleNewSuggestionSubmission}
-              className="w-full p-2 rounded-lg"
-              style={{ backgroundColor: board?.accentColor || "#6366f1", color: board?.secondaryColor || "#fff" }}
-              disabled={submitting}
-            >
-              {submitting ? "Submitting..." : "Submit"}
-            </button>
-          </div>
-          <PoweredByBadge primaryColor={board?.primaryColor} />
+      <div className="md:w-1/3 w-full p-4">
+        <div>
+          {board?.logo ? (
+            <Image src={board.logo} alt={`${board.name} logo`} width={175} height={175} className="mb-4" />
+          ) : (
+            <p className="text-xl font-bold mb-4 w-full break-words">{board?.name}</p>
+          )}
         </div>
-        <div className="w-2/3 p-4">
-          {hideEmptyMessage ? null : <p className="text-md font-semibold">No feedback yet</p>}
-          {board?.suggestions.map((suggestion: Suggestion, index: number) => (
-            <SuggestionCard key={index} suggestion={suggestion} boardData={board} />
-          ))}
+        <div className="p-6 rounded-lg" style={{ backgroundColor: board?.secondaryColor || "#f9fafb" }}>
+          <h1 className="text-lg font-semibold mb-6">Add feedback</h1>
+          <div className="mb-4">
+            <label className="block text-xs font-bold mb-1" htmlFor="suggestionTitle">
+              Post Title
+            </label>
+            <input
+              id="suggestionTitle"
+              type="text"
+              placeholder="Add magic link authentication"
+              className="w-full p-2 rounded-lg focus:outline-none"
+              value={suggestionTitle}
+              onChange={(e) => setsuggestionTitle(e.target.value)}
+              style={{ backgroundColor: lighterSecondaryColor }}
+            />
+          </div>
+          <div className="mb-4">
+            <label className="block text-xs font-bold mb-1" htmlFor="suggestionDescription">
+              Post Text
+            </label>
+            <textarea
+              id="suggestionDescription"
+              placeholder="I hate having to keep track of all these passwords..."
+              className="w-full p-2 rounded-lg focus:outline-none"
+              value={suggestionDescription}
+              onChange={(e) => setsuggestionDescription(e.target.value)}
+              style={{ backgroundColor: lighterSecondaryColor, height: "100px" }}
+            />
+          </div>
           <button
-            onClick={handleLoadMoreSuggestionsSubmission}
-            className={"w-full p-2 rounded-lg" + (hideLoadMoreButton ? " hidden" : "")}
+            onClick={handleNewSuggestionSubmission}
+            className="w-full p-2 rounded-lg"
             style={{ backgroundColor: board?.accentColor || "#6366f1", color: board?.secondaryColor || "#fff" }}
             disabled={submitting}
           >
-            {submitting ? "Loading..." : "Load More"}
+            {submitting ? "Submitting..." : "Submit"}
           </button>
         </div>
+        <PoweredByBadge primaryColor={board?.primaryColor} />
+      </div>
+      <div className="md:w-2/3 w-full p-4">
+        {hideEmptyMessage ? null : <p className="text-md font-semibold">No feedback yet</p>}
+        {board?.suggestions.map((suggestion: Suggestion, index: number) => (
+          <SuggestionCard key={index} suggestion={suggestion} boardData={board} />
+        ))}
+        <button
+          onClick={handleLoadMoreSuggestionsSubmission}
+          className={"w-full p-2 rounded-lg" + (hideLoadMoreButton ? " hidden" : "")}
+          style={{ backgroundColor: board?.accentColor || "#6366f1", color: board?.secondaryColor || "#fff" }}
+          disabled={submitting}
+        >
+          {submitting ? "Loading..." : "Load More"}
+        </button>
       </div>
     </main>
   )
